# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API documentation for Printart Accurate",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "schemes": [
        "http",
        "https"
    ],
    "produces": [
        "application/json",
        "application/xml"
    ],
    "paths": {
...
        "/account-count": {
            "get": {
                "tags": [
                    "GLAccount"
                ],
                "summary": "Count GL Account Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/item-count": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Count Item Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/transaction-count": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Count Transaction Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Find User by Id",
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user that needs to be fetched",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "USERID": {
                                            "type": "number"
                                        },
                                        "USERNAME": {
                                            "type": "string"
                                        },
                                        "USERLEVEL": {
                                            "type": "string"
                                        },
                                        "FULLNAME": {
                                            "type": "string"
                                        },
                                        "USERPASSWORD": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update User",
                "operationId": "edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "Request",
                        "in": "body",
                        "description": "Input request",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RequestUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "USERID": {
                                            "type": "number"
                                        },
                                        "USERNAME": {
                                            "type": "string"
                                        },
                                        "USERLEVEL": {
                                            "type": "string"
                                        },
                                        "FULLNAME": {
                                            "type": "string"
                                        },
                                        "USERPASSWORD": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/users-count": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Count User Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/warehouse-count": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Count Warehouse Data",
                "operationId": "countData",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search - eg: Name",
                        "type": "string"
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Retrieve data by date from - eg: 2019-08-01",
                        "type": "string"
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Retrieve data by date to - eg: 2019-08-01",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Response",
                        "schema": {
                            "property": "data",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                },
                                "data": {
                                    "properties": {
                                        "count": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "422": {
                        "description": "Error Response",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "meta": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "detail": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "definitions": {
...
        "RequestUser": {
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "Muhammad"
                },
                "lastName": {
                    "type": "string",
                    "example": "Al-Pandi"
                },
                "userLevel": {
                    "type": "number",
                    "example": 2
                },
                "username": {
                    "type": "string",
                    "example": "pandi95"
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}